pyproject.toml
[project]
name = "v1"
version = "0.1.0"
description = ""
authors = [
    {name = "faiver90",email = "faiver90@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q --cov=app --cov-report=term-missing --cov-fail-under=70"

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = ["RUF012"]
exclude = [
    "*/migrations/*",  # Все файлы в папке migrations
]
[tool.ruff.lint.per-file-ignores]
"**/settings/*.py" = ["F403", "F405"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_calls = false
disallow_untyped_decorators = false
disallow_incomplete_defs = false
check_untyped_defs = false
exclude = [
    "(?x)(^|/)migrations/.*\\.py$",
]


[tool.black]
line-length = 79